Add logic to add a company to user's favorite list.

+++ app/controllers/api/startup_fundraising/v0/nbos/associates_controller.rb
@@ -4,7 +4,7 @@ class Api::StartupFundraising::V0::Nbos::AssociatesController < Api::StartupFund
  
   
   def index
-    if params[:id].present?
+    if params[:id].present? && @token_details.present? && @token_details.username.present?
     	@company = Com::Nbos::StartupFundraising::Company.where(id: params[:id]).first
     	if @company.present?
     		render :json => @company.company_associates
@@ -17,16 +17,16 @@ class Api::StartupFundraising::V0::Nbos::AssociatesController < Api::StartupFund
   end
 
   def create
-     if params[:id].present?
-       company = Com::Nbos::StartupFundraising::Company.where(id: params[:id]).first
-       if company.present? 
+     if params[:id].present? && @token_details.present? && @token_details.username.present?
+       @company = Com::Nbos::StartupFundraising::Company.where(id: params[:id]).first
+       if @company.present? 
 	       associated_team_type = Com::Nbos::StartupFundraising::AssociateTeam.where(name: params[:associate_type]).first if params[:associate_type].present?
          associate_params = params[:associate].except(:associate_type)
          @team_member = Com::Nbos::StartupFundraising::CompanyAssociate.new(associate_params.permit!)
 	       @team_member.associate_team_id = associated_team_type.present? ? associated_team_type.id : nil
-	       @team_member.company_id = company.id
+	       @team_member.company_id = @company.id
 	       if @team_member.save
-	         render :json => @team_member
+	         render :json => @company.company_associates
 	       else
 	         render :json => {status: 500, message: @team_member.errors.messages}
 	       end
@@ -52,7 +52,7 @@ class Api::StartupFundraising::V0::Nbos::AssociatesController < Api::StartupFund
   end
   
   def update_associate
-  	if params[:id].present?
+  	if params[:id].present? && @token_details.present? && @token_details.username.present?
        @company_associate = Com::Nbos::StartupFundraising::CompanyAssociate.where(id: params[:id]).first
        associate_params = params[:associate].except(:associate_type)
        if params[:associate_type].present? 
@@ -70,10 +70,12 @@ class Api::StartupFundraising::V0::Nbos::AssociatesController < Api::StartupFund
   end
 
   def delete
-  	if params[:id].present?
+  	if params[:id].present? && @token_details.present? && @token_details.username.present?
        @company_associate = Com::Nbos::StartupFundraising::CompanyAssociate.where(id: params[:id]).first
-       if @company_associate.destroy  
-	       render :json => {status: 200, message: "Associate Deleted Successfully"}
+       company_id = @company_associate.company_id if @company_associate.present?
+       if @company_associate.destroy 
+         @company = Com::Nbos::StartupFundraising::Company.where(id: company_id).first 
+	       render :json => @company.company_associates
        else
        	 render :json => {status: 404, message: "Associate Not Found"}
        end
